add a heuristic search optimization to Glyph()
A heuristic search that depends on the previous binary search result.
Since most of the time we're looking for a rune from the
same language, these runes might be very close to each other.

If there are no gaps in between these two runes, we
can find the target rune data index with a simple calculation.

If there is a gap, this strategy will fail and we'll proceed
to the binary search below.

It can give around 20-25% Glyph() speedup
and turns the lookup into O(1).